+++ crates/infrastructure/src/redis/mod.rs
@@ -18,7 +18,6 @@
 pub mod cache;
 pub mod msgbus;
 pub mod queries;
-pub mod transformer;
 
 use std::time::Duration;
